Release 1.0 version of rules.csv parser
Happy 0.98! This project has been in quiet development since mid 0.97, but with the release of 0.98, and the HUGE lore drop, formally hosting this seems helpful.

Includes the python script necessary to create the parsed text, as well as a copy of the parsed rules.csv file for citation purposes.

Future goals will be to better automate and clean up that parsed file, and chase down a few lingering issues with the way the parser reads some of David's more complex interactions.